move collapse button into the tips
