+++ frontend/public/locales/en/translation.json
@@ -980,8 +980,8 @@
   "videos": {
     "dialogs": {
       "tutorial": {
-        "showTips": "Show the tips",
-        "hideTips": "Hide the tips",
+        "showTips": "Show more",
+        "hideTips": "Show less",
         "title1": "Welcome to Tournesol 🌻",
         "message1": {
           "p10": "This tutorial will guide you, step by step, through a series of four comparisons to help you contribute to the Tournesol project.",
